requirements.txt
djangorestframework>=3.15
djangorestframework-simplejwt>=2.8
django-filter>=23.1
django-cors-headers>=4.0 